升级UIKit

+++ app_main/src/main/java/app/itgungnir/kwa/main/tree/TreeFragment.kt
@@ -3,7 +3,6 @@ package app.itgungnir.kwa.main.tree
 import android.widget.TextView
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.RecyclerView
-import app.itgungnir.kwa.common.dp2px
 import app.itgungnir.kwa.common.popToast
 import app.itgungnir.kwa.main.R
 import app.itgungnir.kwa.main.tree.navigation.NavigationDialog
@@ -11,6 +10,7 @@ import app.itgungnir.kwa.main.tree.tools.ToolsDialog
 import kotlinx.android.synthetic.main.fragment_tree.*
 import my.itgungnir.rxmvvm.core.mvvm.BaseFragment
 import my.itgungnir.rxmvvm.core.mvvm.buildActivityViewModel
+import my.itgungnir.ui.dp2px
 import my.itgungnir.ui.easy_adapter.EasyAdapter
 import my.itgungnir.ui.easy_adapter.bind
 import my.itgungnir.ui.status_view.StatusView
@@ -47,7 +47,7 @@ class TreeFragment : BaseFragment() {
             // Status View
             statusView().addDelegate(StatusView.Status.SUCCEED, R.layout.view_status_list) {
                 it.findViewById<RecyclerView>(R.id.list).apply {
-                    bottomPadding = context.dp2px(10.0F)
+                    bottomPadding = context.dp2px(10.0F).toInt()
                     listAdapter = this.bind()
                         .addDelegate({ true }, TreeDelegate())
                         .initialize()
